Rudimentary TwitchTV login support.
Usernames and passwords will not touch our service, the entire login process is passed off to TwitchTV. We are returned a user key tied to the user's account, which is saved in their cookies.
Added jquery-1.9.1 in the javascripts static folder, but it's generally not ideal to use a locally hosted copy for bandwidth reasons. Google's CDN has a copy.